move context holder to internal
