+++ src/main/java/dev/dbos/transact/DBOS.java
@@ -1,7 +1,6 @@
 package dev.dbos.transact;
 
 import dev.dbos.transact.config.DBOSConfig;
-import dev.dbos.transact.context.DBOSContextHolder;
 import dev.dbos.transact.execution.DBOSExecutor;
 import dev.dbos.transact.execution.ThrowingRunnable;
 import dev.dbos.transact.execution.WorkflowFunction;
@@ -9,6 +8,7 @@ import dev.dbos.transact.execution.WorkflowFunctionWrapper;
 import dev.dbos.transact.interceptor.AsyncInvocationHandler;
 import dev.dbos.transact.interceptor.QueueInvocationHandler;
 import dev.dbos.transact.interceptor.UnifiedInvocationHandler;
+import dev.dbos.transact.internal.DBOSContextHolder;
 import dev.dbos.transact.internal.QueueRegistry;
 import dev.dbos.transact.internal.WorkflowRegistry;
 import dev.dbos.transact.migrations.MigrationManager;
